chore: Refactor repository classes removing return type
