+++ src/repositories/prisma/prisma-books-repository.ts
@@ -1,9 +1,8 @@
-import { Book } from '@prisma/client'
 import { BooksRepository } from '../books-repository'
 import { prisma } from '@/repositories/prisma/prisma'
 
 export class PrismaBooksRepository implements BooksRepository {
-  async findAll(): Promise<Book[]> {
+  async findAll() {
     return await prisma.book.findMany()
   }
 }
